refactor(InputNumberValidatorTest): 테스트 로직 병합
- 기존 InputViewTest에서 테스트하는 내용은 검증에 관련된 내용이기 때문에 더 넓은 범주인 InputNumberValidatorTest로 병합
